app.py 및 main.py의 모듈화
