+++ main.py
@@ -1,45 +1,22 @@
-from typing import Optional, List
-from fastapi import FastAPI, File, UploadFile
-import os
-import streamlit as st
-import subprocess
-
-import io
-import os
-import yaml
-import json
-
-import urllib.request
-
+from fastapi import FastAPI
+from service.back_function import *
 from pydantic import BaseModel
-from transformers import BartForConditionalGeneration, AutoTokenizer
-from text2image.model.dalle.models import Rep_Dalle
-from transformers import BartForConditionalGeneration, AutoTokenizer
 
-from text2image.model.dalle.utils.utils import set_seed, clip_score
+# from text2image.model.dalle.models import Rep_Dalle
+from transformers import BartForConditionalGeneration, AutoTokenizer
+# from text2image.model.dalle.utils.utils import set_seed, clip_score
+
 import clip
-from PIL import Image
-import numpy as np
-
-import urllib.request
 import json
-
-import nltk
-from nltk.tokenize import word_tokenize
-from nltk.tag import pos_tag
-from nltk.corpus import stopwords
-nltk.download('stopwords')
-nltk.download('averaged_perceptron_tagger')
-nltk.download('punkt')
-# !necessary : papago api client 정보를 저장한 json file 과 해당 file path
+import numpy as np
 
 # load model
 global model
 model = BartForConditionalGeneration.from_pretrained('chi0/kobart-dial-sum')
 global tokenizer
 tokenizer = AutoTokenizer.from_pretrained('chi0/kobart-dial-sum')
-global txt2imgModel
-txt2imgModel,_ = Rep_Dalle.from_pretrained("text2image/model/tf_model/model/29052022_082436")
+# global txt2imgModel
+# txt2imgModel,_ = Rep_Dalle.from_pretrained("text2image/model/tf_model/model/29052022_082436")
 
 app = FastAPI()
 
@@ -47,13 +24,7 @@ global client
 with open("./client.json", "r") as file:
     client = json.load(file)
 
-
-def postprocess_text_first_sent(preds):
-    preds = [pred.strip() for pred in preds]
-    preds = [pred[:pred.index(".")+1] if "." in pred else pred for pred in preds]
-
-    return preds
-
+################ 요약 ################
 def generate_summary(dialogue:str):
     global model
     global tokenizer
@@ -82,69 +53,7 @@ def generate_summary(dialogue:str):
 
     return output_str
 
-
-################ 번역 ################
-
-
-def mt(sentence, client_id, client_secret):
-    koText = urllib.parse.quote(sentence)
-    data = "source=ko&target=en&text=" + koText
-    url = "https://openapi.naver.com/v1/papago/n2mt"
-    request = urllib.request.Request(url)
-    request.add_header("X-Naver-Client-Id",client_id)
-    request.add_header("X-Naver-Client-Secret",client_secret)
-    response = urllib.request.urlopen(request, data=data.encode("utf-8"))
-    rescode = response.getcode()
-    if(rescode==200):
-        response_body = response.read()
-        json_result = json.loads(response_body.decode('utf-8'))
-        enText = json_result["message"]["result"]["translatedText"]
-        return enText
-    else:
-        print("Error Code:" + rescode)
-
-
-################ 전처리 ################
-
-
-def tokNVJR(sentence):
-    tokenized = []
-    sentence = word_tokenize(sentence)
-    tags = pos_tag(sentence)
-    for (word, tag) in tags:
-        if tag[0]=='N' or tag[0]=='V' or tag[0]=='J' or tag[0]=='R':
-            tokenized.append(word)
-
-    return tokenized
-
-
-def tokSTOP(sentence):
-    sw = stopwords.words('english')
-    sentence = word_tokenize(sentence.lower())
-    words = [word for word in sentence if word not in sw]
-    
-    return words
-
-
-def transformText(text):
-
-    sentences = []
-    sentences.append(", ".join(tokSTOP(text)))
-    sentences.append(", ".join(tokNVJR(text)))
-    return sentences
-
-
-def preprocess(sentence):
-    prefix = "A painting of "
-    answer = []
-    print(transformText(sentence))
-    for sentence in transformText(sentence):
-        answer.append(prefix + sentence)
-    
-    return answer
-
 ################ 번역 + 전처리 ################
-
 def ko2en(sentence):
     global client
     client_id, client_secret = client["client_id"], client["client_secret"]
@@ -152,6 +61,7 @@ def ko2en(sentence):
     sentences = preprocess(sentence)
     return sentences
 
+"""
 ################ Text2Image ################
 
 def txt2img(text):
@@ -198,6 +108,7 @@ def txt2img(text):
     new_im = np.array(new_im)
     new_im = new_im.tolist()
     return new_im
+"""
 
 ################ Text2Image ################
 
@@ -209,8 +120,7 @@ async def upload_image(item: Item):
     kor_sum = postprocess_text_first_sent(generate_summary(item.dialogue))
 
     result = ko2en(kor_sum[0])
-    stop_img = txt2img(result[0])
-    # NVJR_imgs = txt2img(result[1])
+    # stop_img = txt2img(result[0])
 
-
-    return {"summary": result, "kor_sum":kor_sum[0], "image_array":stop_img}
+    # return {"summary": result, "kor_sum":kor_sum[0], "image_array":stop_img}
+    return {"summary": result, "kor_sum":kor_sum[0]}
